chore: unit tests refactored
