+++ tests/spanner_dbapi/test_cursor.py
@@ -105,7 +105,7 @@ class TestCursor(unittest.TestCase):
             "google.cloud.spanner_dbapi.parse_utils.classify_stmt",
             return_value=parse_utils.STMT_DDL,
         ) as mock_classify_stmt:
-            sql = 'sql'
+            sql = "sql"
             cursor.execute(sql=sql)
             mock_classify_stmt.assert_called_once_with(sql)
             self.assertEqual(cursor._connection.ddl_statements, [sql])
@@ -299,66 +299,4 @@ class TestCursor(unittest.TestCase):
 
         execute_mock.assert_has_calls(
             (mock.call(operation, (1,)), mock.call(operation, (2,)))
-        )
-
-
-class TestColumns(unittest.TestCase):
-    def test_ctor(self):
-        from google.cloud.spanner_dbapi.cursor import ColumnInfo
-
-        name = "col-name"
-        type_code = 8
-        display_size = 5
-        internal_size = 10
-        precision = 3
-        scale = None
-        null_ok = False
-
-        cols = ColumnInfo(
-            name,
-            type_code,
-            display_size,
-            internal_size,
-            precision,
-            scale,
-            null_ok,
-        )
-
-        self.assertEqual(cols.name, name)
-        self.assertEqual(cols.type_code, type_code)
-        self.assertEqual(cols.display_size, display_size)
-        self.assertEqual(cols.internal_size, internal_size)
-        self.assertEqual(cols.precision, precision)
-        self.assertEqual(cols.scale, scale)
-        self.assertEqual(cols.null_ok, null_ok)
-        self.assertEqual(
-            cols.fields,
-            (
-                name,
-                type_code,
-                display_size,
-                internal_size,
-                precision,
-                scale,
-                null_ok,
-            ),
-        )
-
-    def test___get_item__(self):
-        from google.cloud.spanner_dbapi.cursor import ColumnInfo
-
-        fields = ("col-name", 8, 5, 10, 3, None, False)
-        cols = ColumnInfo(*fields)
-
-        for i in range(0, 7):
-            self.assertEqual(cols[i], fields[i])
-
-    def test___str__(self):
-        from google.cloud.spanner_dbapi.cursor import ColumnInfo
-
-        cols = ColumnInfo("col-name", 8, None, 10, 3, None, False)
-
-        self.assertEqual(
-            str(cols),
-            "ColumnInfo(name='col-name', type_code=8, internal_size=10, precision='3')",
         )
